SceneLoader and Hand scripts got updated to compensate for the bug where if the hands triggered something at the start of the fade, they do not transition into the new scene. With these additions, the transition should happen. Also added triggers into MoodMatic to be able to go back to Lobby.
SceneLoader and Hand scripts got updated to compensate for the bug where if the hands triggered something at the start of the fade, they do not transition into the new scene. With these additions, the transition should happen. Also added triggers into MoodMatic to be able to go back to Lobby.